@colega 2024-05-21
@jesusvazquez nerd-sniped me, so I'll do the same with the rest, here's the difference in assembly between doing modulo 128 and modulo 100:
https://godbolt.org/z/d86cG8Eej
```asm
        TEXT    main.modulo128(SB), NOSPLIT|NOFRAME|ABIInternal, $0-8
        FUNCDATA        $0, gclocals·g2BeySu+wFnoycgXfElmcg==(SB)
        FUNCDATA        $1, gclocals·g2BeySu+wFnoycgXfElmcg==(SB)
        FUNCDATA        $5, main.modulo128.arginfo1(SB)
        FUNCDATA        $6, main.modulo128.argliveinfo(SB)
        PCDATA  $3, $1
        MOVQ    AX, CX
        SARQ    $63, CX
        SHRQ    $57, CX
        ADDQ    AX, CX
        ANDQ    $-128, CX
        SUBQ    CX, AX
        RET
        TEXT    main.modulo100(SB), NOSPLIT|NOFRAME|ABIInternal, $0-8
        FUNCDATA        $0, gclocals·g2BeySu+wFnoycgXfElmcg==(SB)
        FUNCDATA        $1, gclocals·g2BeySu+wFnoycgXfElmcg==(SB)
        FUNCDATA        $5, main.modulo100.arginfo1(SB)
        FUNCDATA        $6, main.modulo100.argliveinfo(SB)
        PCDATA  $3, $1
        MOVQ    AX, CX
        SARQ    $63, AX
        MOVQ    AX, DX
        MOVQ    $-6640827866535438581, AX
        MOVQ    DX, BX
        IMULQ   CX
        ADDQ    CX, DX
        SARQ    $6, DX
        SUBQ    BX, DX
        IMUL3Q  $100, DX, DX
        SUBQ    DX, CX
        MOVQ    CX, AX
        RET
```