@cristaloleg 2024-05-26
Small clarifications: there is no need to explicitly rewrite `a % 128` as `a & 127` 'cause Go compiler does this automatically, see `src/cmd/compile/internal/ssa/_gen/generic.rules` (look for `mod by power of 2 constant.`).